Fix int32 overflow on filesize_approx for files >2GB
DownloadInfo.Size was int32, so a filesize_approx above ~2.1GB (e.g. long
high-bitrate mixes) failed JSON unmarshal with "cannot unmarshal number ...
into Go struct field DownloadInfo.filesize_approx of type int32" and the
download never started. Widened to int64.

+++ server/common/types.go
@@ -8,7 +8,7 @@ type DownloadInfo struct {
 	Title       string    `json:"title"`
 	Thumbnail   string    `json:"thumbnail"`
 	Resolution  string    `json:"resolution"`
-	Size        int32     `json:"filesize_approx"`
+	Size        int64     `json:"filesize_approx"`
 	VCodec      string    `json:"vcodec"`
 	ACodec      string    `json:"acodec"`
 	Extension   string    `json:"ext"`
